Move TimeProvider to an infrastructure:time module instead of being in util.
