+++ infrastructure/time/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+  api project(':infrastructure:events')
+
+  testFixturesImplementation project(':infrastructure:unsigned')
+}
