+++ infrastructure/time/src/main/java/tech/pegasys/teku/infrastructure/time/TimeProvider.java
@@ -11,21 +11,17 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.time;
+package tech.pegasys.teku.infrastructure.time;
 
-import java.time.Clock;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
-public class SystemTimeProvider implements TimeProvider {
+public interface TimeProvider {
 
-  private final Clock clock;
+  UInt64 MILLIS_PER_SECOND = UInt64.valueOf(1000);
 
-  public SystemTimeProvider() {
-    this.clock = Clock.systemUTC();
-  }
+  UInt64 getTimeInMillis();
 
-  @Override
-  public UInt64 getTimeInMillis() {
-    return UInt64.valueOf(clock.millis());
+  default UInt64 getTimeInSeconds() {
+    return getTimeInMillis().dividedBy(MILLIS_PER_SECOND);
   }
 }
