+++ networking/eth2/build.gradle
@@ -31,6 +31,7 @@ dependencies {
   testImplementation testFixtures(project(':ethereum:datastructures'))
   testImplementation testFixtures(project(':ethereum:statetransition'))
   testImplementation testFixtures(project(':infrastructure:async'))
+  testImplementation testFixtures(project(':infrastructure:time'))
   testImplementation testFixtures(project(':networking:p2p'))
   testImplementation testFixtures(project(':util'))
   testImplementation testFixtures(project(':storage'))
@@ -50,6 +51,7 @@ dependencies {
   testFixturesImplementation testFixtures(project(':infrastructure:async'))
   testFixturesImplementation testFixtures(project(':util'))
   testFixturesImplementation testFixtures(project(':infrastructure:events'))
+  testFixturesImplementation testFixtures(project(':infrastructure:time'))
   testFixturesImplementation testFixtures(project(':storage'))
   testFixturesImplementation project(':infrastructure:subscribers')
   testFixturesImplementation project(':services:serviceutils')
