+++ pow/build.gradle
@@ -17,6 +17,7 @@ dependencies {
 
   testImplementation testFixtures(project(':util'))
   testImplementation testFixtures(project(':infrastructure:async'))
+  testImplementation testFixtures(project(':infrastructure:time'))
   testImplementation testFixtures(project(':ethereum:datastructures'))
 
   testFixturesImplementation project(':util')
