+++ pow/src/main/java/tech/pegasys/teku/pow/Eth1BlockFetcher.java
@@ -26,7 +26,7 @@ import org.web3j.protocol.core.methods.response.EthBlock.Block;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.pow.api.Eth1EventsChannel;
-import tech.pegasys.teku.util.time.TimeProvider;
+import tech.pegasys.teku.infrastructure.time.TimeProvider;
 
 public class Eth1BlockFetcher {
   private static final Logger LOG = LogManager.getLogger();
