+++ pow/src/test/java/tech/pegasys/teku/pow/Eth1StatusLoggerTest.java
@@ -19,7 +19,7 @@ import static org.mockito.Mockito.mock;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.infrastructure.async.StubAsyncRunner;
-import tech.pegasys.teku.util.time.TimeProvider;
+import tech.pegasys.teku.infrastructure.time.TimeProvider;
 
 class Eth1StatusLoggerTest {
