+++ settings.gradle
@@ -24,6 +24,7 @@ include 'infrastructure:metrics'
 include 'infrastructure:unsigned'
 include 'infrastructure:logging'
 include 'infrastructure:subscribers'
+include 'infrastructure:time'
 include 'networking:eth2'
 include 'networking:p2p'
 include 'services'
