+++ sync/build.gradle
@@ -24,6 +24,7 @@ dependencies {
     testImplementation testFixtures(project(':ethereum:statetransition'))
     testImplementation project(':ethereum:weaksubjectivity')
     testImplementation testFixtures(project(':infrastructure:async'))
+    testImplementation testFixtures(project(':infrastructure:time'))
     testImplementation testFixtures(project('::networking:eth2'))
     testImplementation testFixtures(project('::networking:p2p'))
     testImplementation testFixtures(project(':util'))
