+++ sync/src/main/java/tech/pegasys/teku/sync/multipeer/chains/SyncSourceFactory.java
@@ -21,7 +21,7 @@ import java.util.Map;
 import tech.pegasys.teku.infrastructure.async.AsyncRunner;
 import tech.pegasys.teku.networking.eth2.peers.Eth2Peer;
 import tech.pegasys.teku.networking.eth2.peers.SyncSource;
-import tech.pegasys.teku.util.time.TimeProvider;
+import tech.pegasys.teku.infrastructure.time.TimeProvider;
 
 public class SyncSourceFactory {
