+++ teku/src/main/java/tech/pegasys/teku/BeaconNode.java
@@ -38,7 +38,7 @@ import tech.pegasys.teku.services.BeaconNodeServiceController;
 import tech.pegasys.teku.util.cli.VersionProvider;
 import tech.pegasys.teku.util.config.Constants;
 import tech.pegasys.teku.util.config.GlobalConfiguration;
-import tech.pegasys.teku.util.time.SystemTimeProvider;
+import tech.pegasys.teku.infrastructure.time.SystemTimeProvider;
 
 public class BeaconNode implements Node {
