+++ util/build.gradle
@@ -6,6 +6,7 @@ dependencies {
   api project(':infrastructure:crypto')
   api project(':infrastructure:metrics')
   api project(':infrastructure:events')
+  api project(':infrastructure:time')
 
   implementation project(':bls')
   implementation project(':infrastructure:async')
